+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Added test suite to test wallet
 - Added get balance method to fetch STX balance of an account and test cases
 - Added functionality to sign and send native STX token, fungible tokens
+- Added functionality to sign message